fix: externals not being ignored
