+++ tests/fixtures/external.ts
@@ -1,0 +1,1 @@
+import "mapped";
